WIP Strucutre of array of dicts now correct
Previous bug almost 100% fixed where child attribs
were sometimes stored incorrectly. There were instances where it
should've looked like [{},{},{}] in the resulting json, but was looking
like {{},{},{}}. This is fixed except for block items, which might need
to be hardocded, or was an error on the original author's attempt

Next work:
1. Not exploring the full depth of a node (compound's block items are
not all being explored)
2. Missing attributes -- there are cases where when an attribute is not filled
in the AST, it should be marked as NULL in the json, but for right now my approach
is just omitting these entries from the json.