Serialization of AST to JSON complete
This serialization allows the transformation of the dict
representation of the AST (from pycparser) to json representation.

There are notably 2 hardocded things that might seem weird, and are
still weird to me. 'ext' and 'block_items' in the dict representation
are always lists, even whenever they have only one item. This is inconsistent
with the rest of the nodes that will appear in AST traversal.

The usualy approach is if the child attribute has only 1 child, then the
child attribute should be represented as a dictionary in the json. The counter
part is that if the child attribute has >1 child, the child attribute should
be represetned as an array of dicts. For whatever reason, ext and block_items
are the exception to this rule.